♻️ Refactor CDN work (#708) * Move S3-related code into a service within `cms/src/services` * Move image migration script to `/scripts`, use pm2 config instead of process.env
