+++ cms/pm2.config.js
@@ -40,6 +40,9 @@ module.exports = {
         ES_API_VERSION: 6.1,
         ES_LOG_LEVEL: 'error',
         AUTH_ENABLED: false,
+        S3_BUCKET: 'hcmi-demo',
+        IAM_USER_KEY: 'ADD-KEY',
+        IAM_USER_SECRET: 'ADD-SECRET',
       },
     },
   ],
