+++ cms/src/services/s3/index.js
@@ -1,0 +1,1 @@
+export { uploadToS3, deleteFromS3 } from './s3';
